deploy_onnx/requirements.txt
fastapi
uvicorn[standard]
python-multipart
onnxruntime
Pillow
gunicorn